transaction bug fixed

+++ src/main/java/org/flow/controllers/UserController.java
@@ -38,8 +38,12 @@ public class UserController {
 
     //get all users
     @GetMapping
-    public @ResponseBody ResponseEntity getAllUsers () {
-        return ResponseEntity.ok(userRepository.findAllByOrderByLastNameAsc());
+    public @ResponseBody ResponseEntity getAllUsers (@RequestHeader(value = "Authorization") String token) {
+        if(validations.stayingALive(token)) {
+            return ResponseEntity.ok(userRepository.findAllByOrderByLastNameAsc());
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Session validations.");
+        }
     }
 
     //get user by ID
@@ -64,33 +68,37 @@ public class UserController {
 
     //create new user
     @PostMapping
-    public @ResponseBody ResponseEntity addNewUser (@RequestBody String user) {
-        User newUser = new User();
-        JSONObject jsonObject = new JSONObject(user);
-        if(userRepository.findByEmail(jsonObject.getString("email")) != null) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("This e-mail address is already taken.");
-        } else {
-            PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-            String hashedPassword = passwordEncoder.encode(jsonObject.getString("password"));
-            newUser.setPassword(hashedPassword);
-            newUser.setFirstName(jsonObject.getString("firstName"));
-            newUser.setLastName(jsonObject.getString("lastName"));
-            newUser.setNickName(jsonObject.getString("nickName"));
-            newUser.setEmail(jsonObject.getString("email"));
-            newUser.setRoleType(jsonObject.getString("role"));
-            if (jsonObject.getInt("xp") < 0) {
-                newUser.setXp(jsonObject.getInt("xp"));
+    public @ResponseBody ResponseEntity addNewUser (@RequestBody String user, @RequestHeader(value = "Authorization") String token) {
+        if(validations.stayingALive(token)) {
+            User newUser = new User();
+            JSONObject jsonObject = new JSONObject(user);
+            if (userRepository.findByEmail(jsonObject.getString("email")) != null) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("This e-mail address is already taken.");
+            } else {
+                PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+                String hashedPassword = passwordEncoder.encode(jsonObject.getString("password"));
+                newUser.setPassword(hashedPassword);
+                newUser.setFirstName(jsonObject.getString("firstName"));
+                newUser.setLastName(jsonObject.getString("lastName"));
+                newUser.setNickName(jsonObject.getString("nickName"));
+                newUser.setEmail(jsonObject.getString("email"));
+                newUser.setRoleType(jsonObject.getString("role"));
+                if (jsonObject.getInt("xp") < 0) {
+                    newUser.setXp(jsonObject.getInt("xp"));
+                }
+                Date dob = null;
+                try {
+                    dob = new SimpleDateFormat("yyyy-MM-dd").parse(jsonObject.getString("dob"));
+                } catch (ParseException e) {
+                    e.printStackTrace();
+                }
+                newUser.setDob(dob);
+                newUser.setGender(jsonObject.getBoolean("gender"));
+                userRepository.save(newUser);
+                return ResponseEntity.ok(newUser);
             }
-            Date dob = null;
-            try {
-                dob = new SimpleDateFormat("yyyy-MM-dd").parse(jsonObject.getString("dob"));
-            } catch (ParseException e) {
-                e.printStackTrace();
-            }
-            newUser.setDob(dob);
-            newUser.setGender(jsonObject.getBoolean("gender"));
-            userRepository.save(newUser);
-            return ResponseEntity.ok(newUser);
+        } else  {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Session validations.");
         }
     }
 
@@ -135,12 +143,15 @@ public class UserController {
     //delete user by ID
     @DeleteMapping(path = "/{id}")
     public @ResponseBody ResponseEntity deleteUser (@PathVariable("id") Long id, @RequestHeader(value = "Authorization")  String token) {
-        if(validations.checkUser(id, token)) {
-            userRepository.deleteById(id);
-            return ResponseEntity.ok(userRepository.findAll());
+        if(validations.stayingALive(token)) {
+            if (validations.checkUser(id, token)) {
+                userRepository.deleteById(id);
+                return ResponseEntity.ok(userRepository.findAll());
+            } else {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("You shall not pass.");
+            }
         } else {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("You shall not pass.");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Session validations.");
         }
-
     }
 }
